Hold open the console slave device in the supervisor
Assaf Gordon <assafgordon@gmail.com> reported odd behaviour with a
contained process which generates output and immediately exits. The
ensuing discussion also revealed further misbehaviour in cases where the
container closes the console and later reopens it using /dev/console.

Without relying on ambiguously-specified behaviour, it is difficult to
correctly construct a poll() + read() loop for a pseudo-terminal master
when the slave device might be opened and closed multiple times.

Fortunately, it is possible to bypass the tricky edge-cases and simplify
the console supervisor by holding the slave device open ourselves. In
this case, poll() will always block unless there is data to read and
there are no EOF conditions to handle.