Make supervisor processes dumpable after privileges are dropped
On recent kernels, one process can only join the user namespace of
another if it has permission to dump or ptrace it.

To prevent an unprivileged user from tampering with the execution of a
more privileged process, the kernel disables the dumpable flag when a
program is run setuid.

Once contain has set the user namespace mappings, it fully drops its
privileges. At this point, it is safe to re-enable the dumpable flag so
that processes can be injected into the container by the invoking user.