Re-exec from a sealed memfd to protect against /proc/self/exe abuse
Linux treats /proc/self/exe and /proc/N/exe as magic. Although stat()
sees a symlink to the absolute path of the binary, open() accesses the
binary itself whether or not the symlink can be resolved in the
filesystem namespace of the opening process.

Inside a container, this leaks a path to any host binary that joins that
container's namespaces. CVE-2019-5736 describes an attack on runC
'privileged containers' using this route: as the host root user is
mapped within the container, a container process can open the relevant
/proc/N/exe file for writing and compromise it.

Unless explicitly requested, contain will never map the host root user
into the container's user namespace, so this shouldn't affect normal
users of contain or other 'unprivileged' supervisors. However, it is
safer to avoid exposing host binaries at all.

The contain binary is always running outside the container as a
supervisor when processes exist inside it. Even if host root is mapped,
attempts to write to contain via /proc/N/exe will fail with ETXTBSY.

However, inject will exec or exit shortly after joining the container.
If the root user in a malicious container had write permission to host
inject, it would be sufficient for an attacker to open the relevant
/proc/N/exe with O_PATH, wait for inject to exec or exit, then re-open
that file descriptor O_WRONLY when the binary is no longer busy.

To protect inject against this, create a new memfd on invocation, clone
the running program to it, seal it to prevent any further modifications,
then re-exec from the memfd clone. /proc/self/exe now refers to an
imutable in-memory copy of inject, not the host binary.

This precuation may later become unnecessary if the kernel handling of
/proc/N/exe pointing outside the current filesystem namespace improves.